Improve parser state for special tokens.
Currently, the onlineParser doesn't alter token state for special tokens like `comment` or `invalidchar`. When you load the token state for one of these tokens, it appears to be whatever came before it. When building "intelisense" features for IDEs this can be a complicating factor.

This proposes adding two new special parse rules "Comment" and "Invalid" and altering the parser behavior to detect and insert them when necessary.

This also results in onlineParser becoming more generic with respect to Comment lexer rules. Previously onlineParser knew more than it should about how comments are defined. Now, the provided parse rules (specifically the lexer rules) supply the definition for `Comment`.

Finally, this removes global state from the parser when saving and resuming backup states, keeping that context contained within the parser function. That removes the possibility for poisoning the parse state.