[RFC] Info tooltips and Jump to definition.
**Info**

When hovering over a token in the query, produces a tooltip with some information about the thing you're hovering over.

This can be really helpful when you're reading through a query and trying to figure out what the current type is, what type an argument accepts, or descriptions of any of those things.

**Jump**

When holding down Command (Mac) or Control (Win) then tokens in the editor become clickable links. Clicking a link can then call a function with relevant schema data referenced, so the editor can do something useful. For example, in GraphiQL we can automatically expose the Documentation explorer to show more about it.